Relocate 4 presenters
CreateCooperationPresenter, GetPlanSummaryCompanyPresenter, GetPlanSummaryMemberPresenter, GetPlanSummaryAccountantPresenter
