+++ arbeitszeit_flask/company/routes.py
@@ -88,7 +88,6 @@ from arbeitszeit_web.create_draft import (
     GetPrefilledDraftDataPresenter,
 )
 from arbeitszeit_web.get_company_transactions import GetCompanyTransactionsPresenter
-from arbeitszeit_web.get_plan_summary_company import GetPlanSummaryCompanyPresenter
 from arbeitszeit_web.get_statistics import GetStatisticsPresenter
 from arbeitszeit_web.presenters.company_purchases_presenter import (
     CompanyPurchasesPresenter,
@@ -103,6 +102,9 @@ from arbeitszeit_web.presenters.get_company_summary_presenter import (
 from arbeitszeit_web.presenters.get_coop_summary_presenter import (
     GetCoopSummarySuccessPresenter,
 )
+from arbeitszeit_web.presenters.get_plan_summary_company_presenter import (
+    GetPlanSummaryCompanyPresenter,
+)
 from arbeitszeit_web.presenters.hide_plan_presenter import HidePlanPresenter
 from arbeitszeit_web.presenters.list_all_cooperations_presenter import (
     ListAllCooperationsPresenter,
@@ -112,6 +114,7 @@ from arbeitszeit_web.presenters.query_companies_presenter import QueryCompaniesP
 from arbeitszeit_web.presenters.show_a_account_details_presenter import (
     ShowAAccountDetailsPresenter,
 )
+from arbeitszeit_web.presenters.show_my_plans_presenter import ShowMyPlansPresenter
 from arbeitszeit_web.presenters.show_p_account_details_presenter import (
     ShowPAccountDetailsPresenter,
 )
@@ -127,7 +130,6 @@ from arbeitszeit_web.request_cooperation import (
     RequestCooperationPresenter,
 )
 from arbeitszeit_web.show_my_cooperations import ShowMyCooperationsPresenter
-from arbeitszeit_web.show_my_plans import ShowMyPlansPresenter
 from arbeitszeit_web.url_index import UrlIndex
 
 from .blueprint import CompanyRoute
