+++ arbeitszeit_flask/views/create_cooperation_view.py
@@ -10,7 +10,9 @@ from arbeitszeit.use_cases.create_cooperation import (
 from arbeitszeit_flask.flask_session import FlaskSession
 from arbeitszeit_flask.forms import CreateCooperationForm
 from arbeitszeit_flask.template import UserTemplateRenderer
-from arbeitszeit_web.create_cooperation import CreateCooperationPresenter
+from arbeitszeit_web.presenters.create_cooperation_presenter import (
+    CreateCooperationPresenter,
+)
 
 
 @dataclass
