+++ arbeitszeit_web/presenters/create_cooperation_presenter.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass
 
 from arbeitszeit.use_cases.create_cooperation import CreateCooperationResponse
 
-from .notification import Notifier
-from .translator import Translator
+from ..notification import Notifier
+from ..translator import Translator
 
 
 @dataclass
