+++ tests/presenters/test_get_plan_summary_company_presenter.py
@@ -5,7 +5,9 @@ some functionalities are tested in tests/presenters/test_plan_summary_formatter.
 from uuid import uuid4
 
 from arbeitszeit.use_cases.get_plan_summary import GetPlanSummaryUseCase
-from arbeitszeit_web.get_plan_summary_company import GetPlanSummaryCompanyPresenter
+from arbeitszeit_web.presenters.get_plan_summary_company_presenter import (
+    GetPlanSummaryCompanyPresenter,
+)
 from tests.presenters.base_test_case import BaseTestCase
 from tests.presenters.data_generators import PlanSummaryGenerator
 from tests.session import FakeSession
